% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixed_mult_logit_term_grad}
\alias{mixed_mult_logit_term_grad}
\title{Computes the Gradient}
\usage{
mixed_mult_logit_term_grad(
  eta,
  Sigma,
  which_category,
  weights,
  nodes,
  target_size = 128L,
  use_adaptive = TRUE
)
}
\description{
Computes the gradient of the expectation of the multinomial logit factors
with respect to the offsets on the linear predictor scale.
}
